server/pyproject.toml
[project]
name = "server"
requires-python = ">=3.9, <4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-ai-slim[openai] (>=0.0.21,<0.0.22)",
    "fastapi[standard] (>=0.115.8,<0.116.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


